requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==2.0.12
click==8.1.7
distlib==0.3.8
dj-static==0.0.6
Django==5.0.3
django-apscheduler==0.6.0
django-jazzmin==2.6.0
filelock==3.13.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
idna==3.6
pipenv==2022.5.2
platformdirs==4.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2024.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
tornado==6.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.28.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
yandex-geocoder==2.0.0
zope.event==5.0
zope.interface==6.2
